pyproject.toml
[project]
name = "medcalc-bench"
version = "0.1.0"
description = "MedCalc-Bench: A medical calculator benchmark for LLMs"
requires-python = ">=3.12"
dependencies = [
    "accelerate==0.31.0",
    "aiohttp==3.9.5",
    "beautifulsoup4==4.12.3",
    "click==8.1.7",
    "colorlog==6.8.2",
    "datasets==2.19.2",
    "dill==0.3.8",
    "filelock==3.14.0",
    "fsspec==2024.3.1",
    "huggingface-hub==0.23.3",
    "ipykernel==6.29.4",
    "ipython==8.23.0",
    "ipywidgets==8.1.2",
    "jinja2==3.1.3",
    "joblib==1.4.2",
    "jupyterlab==4.1.6",
    "matplotlib==3.8.4",
    "nltk==3.8.1",
    "numpy==1.26.4",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "openai-parallel-toolkit==0.5.1",
    "openpyxl==3.1.2",
    "packaging==24.0",
    "pandas==2.2.2",
    "pillow==10.3.0",
    "psutil==5.9.8",
    "pyarrow==16.1.0",
    "pydantic==2.7.0",
    "pygments==2.17.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "regex==2024.4.28",
    "requests==2.32.3",
    "rich==13.7.1",
    "safetensors==0.4.3",
    "sentencepiece==0.2.0",
    "simplejson==3.19.2",
    "sympy==1.12",
    "tabulate==0.9.0",
    "tiktoken==0.6.0",
    "together==1.2.0",
    "tokenizers==0.19.1",
    "torch==2.3.0",
    "tqdm==4.66.2",
    "transformers==4.41.2",
    "typer==0.12.3",
    "typing-extensions==4.11.0",
    "urllib3==2.2.1",
]

[dependency-groups]
dev = []
